+++ flow.go
@@ -10,6 +10,18 @@ import (
 	"sort"
 	"strings"
 	"text/tabwriter"
+
+	"github.com/goyek/goyek/v3/internal"
+)
+
+var (
+	signalNotify = signal.Notify
+	signalStop   = signal.Stop
+	osExit       = os.Exit
+
+	// for testing
+	trapSignalsHook       func()
+	trapSignalsSecondHook func()
 )
 
 // Flow is the root type of the package.
@@ -377,7 +389,7 @@ const (
 // Main runs provided tasks and all their dependencies.
 // Each task is executed at most once.
 // It exits the current program when after the run is finished
-// or SIGINT interrupted the execution.
+// or a termination signal interrupted the execution.
 //   - 0 exit code means that non of the tasks failed.
 //   - 1 exit code means that a task has failed or the execution was interrupted.
 //   - 2 exit code means that the input was invalid.
@@ -390,40 +402,86 @@ func Main(args []string, opts ...Option) {
 // Main runs provided tasks and all their dependencies.
 // Each task is executed at most once.
 // It exits the current program when after the run is finished
-// or SIGINT interrupted the execution.
+// or a termination signal interrupted the execution.
 //   - 0 exit code means that non of the tasks failed.
 //   - 1 exit code means that a task has failed or the execution was interrupted.
 //   - 2 exit code means that the input was invalid.
 //
 // Calls [Usage] when invalid args are provided.
 func (f *Flow) Main(args []string, opts ...Option) {
+	osExit(f.runMain(args, opts...))
+}
+
+func (f *Flow) runMain(args []string, opts ...Option) int {
 	out := f.Output()
+	// ensure that the output is synchronized
+	// it is restored in the deferred function
+	// so that the next run of Main or Execute is not affected
+	oldOutput := f.output
+	f.output = internal.SyncWriter(out)
+	defer func() {
+		f.output = oldOutput
+	}()
 
 	// trap Ctrl+C and call cancel on the context
 	ctx, cancel := context.WithCancel(context.Background())
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, os.Interrupt)
-	go func() {
-		<-c // first signal, cancel context
-		fmt.Fprintln(out, "first interrupt, graceful stop")
-		cancel()
+	defer cancel()
 
-		<-c // second signal, hard exit
-		fmt.Fprintln(out, "second interrupt, exit")
-		os.Exit(exitCodeFail)
+	handlerDone := make(chan struct{})
+	handlerFinished := make(chan struct{})
+	go func() {
+		defer close(handlerFinished)
+
+		sigChan := make(chan os.Signal, 1)
+		signalNotify(sigChan, internal.TerminationSignals...)
+		defer signalStop(sigChan)
+
+		// first signal, cancel context
+		if trapSignalsHook != nil {
+			trapSignalsHook()
+		}
+		select {
+		case <-sigChan:
+			fmt.Fprintln(f.output, "first interrupt, graceful stop")
+			cancel()
+		case <-handlerDone:
+			return
+		}
+
+		// second signal, hard exit
+		if trapSignalsSecondHook != nil {
+			trapSignalsSecondHook()
+		}
+		select {
+		case <-sigChan:
+			fmt.Fprintln(f.output, "second interrupt, exit")
+			osExit(exitCodeFail)
+		case <-handlerDone:
+			// finish consuming signals
+			for {
+				select {
+				case <-sigChan:
+					fmt.Fprintln(f.output, "additional interrupt ignored during cleanup")
+				case <-handlerDone:
+					return
+				}
+			}
+		}
 	}()
 
 	exitCode := f.main(ctx, args, opts...)
-	os.Exit(exitCode)
+	close(handlerDone)
+	<-handlerFinished
+	return exitCode
 }
 
 func (f *Flow) main(ctx context.Context, args []string, opts ...Option) int {
 	err := f.Execute(ctx, args, opts...)
-	var ferr *FailError
-	if errors.As(err, &ferr) {
+	if ctx.Err() != nil {
 		return exitCodeFail
 	}
-	if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+	var ferr *FailError
+	if errors.As(err, &ferr) {
 		return exitCodeFail
 	}
 	if err != nil {