+++ internal/signals_default.go
@@ -1,0 +1,9 @@
+// +build !aix,!darwin,!dragonfly,!freebsd,!linux,!netbsd,!openbsd,!solaris
+
+package internal
+
+import "os"
+
+func init() {
+	TerminationSignals = []os.Signal{os.Interrupt}
+}